fix: websocket in dashboard not main
